requirements.txt
python-binance==1.0.30
python-dotenv==1.0.0
rich==13.7.1
pytest==7.4.3
black==23.12.1
flake8==6.1.0
isort==5.13.2
